Rename vendored hdf5 reader to hdf5_example
Avoids colliding with the real daft.datasets.hdf5 reader shipping in Daft soon,
and makes the conversion example self-contained (imports the local copy).
