+++ datasets/egodex/hdf5_example.py
